+++ src/system/windows/system.c
@@ -12,6 +12,7 @@
 //   ZettaScale Zenoh Team, <zenoh@zettascale.tech>
 //
 
+#include <sys/types.h>
 #include <windows.h>
 // The following includes must come after winsock2
 #include <ntsecapi.h>
@@ -46,7 +47,7 @@ uint64_t z_random_u64(void) {
     return ret;
 }
 
-void z_random_fill(void *buf, size_t len) { RtlGenRandom(buf, len); }
+void z_random_fill(void *buf, size_t len) { RtlGenRandom(buf, (ulong)len); }
 
 /*------------------ Memory ------------------*/
 // #define MALLOC(x) HeapAlloc(GetProcessHeap(), 0, (x))
